requirements.txt
loguru
click
numpy
pyzmq
opencv-python 
